Add `-e/--end` to display the target datetime of an unpaused countdown
Motivation:

One might create a countdown for a specific duration, but still be
interested in knowing when the countdown will be over.

For example, in the morning you might create an 8hr timer for work,
pause it while eating lunch, again while walking the dog, and then get
asked by your partner around 16:17 when you are going to stop working
and pay attention to them.

Limitation:

This implementation doesn't show the target incrementing while the
countdown is paused.  However, as soon as the countdown is unpaused,
target will become accurate.