Add `-D/--date-format` to configure the human-readable date format
Motivation:

A countdown might not end today.  To display the target datetime of a
countdown, we must know how to format both the date and time component
of the target in a human-readable way.

`-Z/--time-format` already exists and can be used to configure how to
format the time component; this commit adds the ability to configure
how to format the date component.